Reformat code

+++ server/src/main/java/com/ctof/server/service/impl/UserServiceImpl.java
@@ -3,20 +3,19 @@ package com.ctof.server.service.impl;
 import com.ctof.api.CreateUserRequest;
 import com.ctof.api.UpdateUserRequest;
 import com.ctof.api.User;
+import com.ctof.server.exception.NotFoundException;
 import com.ctof.server.model.UserModel;
 import com.ctof.server.repository.UserRepository;
 import com.ctof.server.service.UserService;
+import org.hibernate.annotations.NotFound;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.List;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -27,15 +26,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Page<User> list(Pageable pageable) {
-        Page<UserModel> userModels =  repo.findAll(pageable);
-        Type pageType = new TypeToken<Page<User>>() {}.getType();
+        Page<UserModel> userModels = repo.findAll(pageable);
+        Type pageType = new TypeToken<Page<User>>() {
+        }.getType();
         return mapper.map(userModels, pageType);
     }
 
     @Override
     public Page<User> listByStartupId(Long startupId, Pageable pageable) {
-        Page<UserModel> userModels =  repo.findByStartupId(startupId, pageable);
-        Type pageType = new TypeToken<Page<User>>() {}.getType();
+        Page<UserModel> userModels = repo.findByStartupId(startupId, pageable);
+        Type pageType = new TypeToken<Page<User>>() {
+        }.getType();
         return mapper.map(userModels, pageType);
     }
 
@@ -49,12 +50,19 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User get(Long id) {
+        UserModel userModel = repo.findOne(id);
+        if (userModel == null) {
+            throw new NotFoundException();
+        }
         return mapper.map(repo.findOne(id), User.class);
     }
 
     @Override
     public User update(Long id, UpdateUserRequest request) {
         UserModel userModel = repo.findOne(id);
+        if (userModel == null) {
+            throw new NotFoundException();
+        }
         mapper.map(request, userModel);
         userModel = repo.save(userModel);
         return mapper.map(userModel, User.class);
@@ -62,6 +70,10 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long id) {
+        UserModel userModel = repo.findOne(id);
+        if (userModel == null) {
+            throw new NotFoundException();
+        }
         repo.delete(id);
     }
 }
